[FT07] - Refatorações Gerais
User Stories: Nenhum
Tarefas: Refatorações em recursos existentes.
Bugs: Nenhum.
Outras Informações: Adicionado funcionalidade para que o piloto de estudo só seja deletado caso não tenha associação com usuários.

+++ src/application/service/pilot.study.service.ts
@@ -64,6 +64,12 @@ export class PilotStudyService implements IPilotStudyService {
     public async remove(id: string): Promise<boolean> {
         try {
             ObjectIdValidator.validate(id)
+            const associatedHealths: number = await this._pilotStudyRepository.countHealthProfessionalsFromPilotStudy(id)
+            const associatedPatients: number = await this._pilotStudyRepository.countPatientsFromPilotStudy(id)
+
+            if (associatedHealths || associatedPatients) {
+                throw new ValidationException(Strings.PILOT_STUDY.HAS_ASSOCIATION)
+            }
             return this._pilotStudyRepository.delete(id)
         } catch (err) {
             return Promise.reject(err)
